streamlit-app/requirements.txt
streamlit==1.35.0
altair==5.3.0
pillow==10.3.0
